refactored complex terms
