+++ ontology/update_complex_terms_worker.ts
@@ -1,8 +1,14 @@
-import { extract, transform } from './migrate_complex_terms_table'
+import {
+	CREATE_COMPLEX_TERMS_TABLE_SQL,
+	extract,
+	INSERT_COMPLEX_TERMS_SQL,
+	transform,
+	type ComplexTerm,
+} from './complex_terms'
 
 export default {
 	// https://developers.cloudflare.com/workers/runtime-apis/handlers/scheduled/
-	async scheduled(event: ScheduledEvent, env: Env, ctx: ExecutionContext) {
+	async scheduled(event: ScheduledEvent, env: Env) {
 		const extracted_rows = await extract()
 
 		if (extracted_rows.length < 1) {
@@ -17,24 +23,12 @@ export default {
 		const num_terms = await env.DB_Ontology.prepare('SELECT COUNT(*) AS count FROM Complex_Terms').first('count')
 		console.log(`updated ${num_terms} complex terms in database.`)
 
-		async function load(terms: ReturnType<typeof transform>) {
+		async function load(terms: ComplexTerm[]) {
 			// https://developers.cloudflare.com/d1/build-databases/query-databases
-			env.DB_Ontology.prepare(`
-				CREATE TABLE IF NOT EXISTS Complex_Terms (
-					'stem' 				TEXT,
-					'sense'				TEXT,
-					'part_of_speech' 	TEXT,
-					'structure'		 	TEXT,
-					'pairing' 			TEXT,
-					'explication' 		TEXT,
-					'ontology_status'	TEXT,
-					'level'				NUMBER,
-					'notes'				TEXT
-				)
-			`).run()
+			env.DB_Ontology.prepare(CREATE_COMPLEX_TERMS_TABLE_SQL).run()
 
 			const clear_stmt = env.DB_Ontology.prepare('DELETE FROM Complex_Terms')
-			const insert_stmt = env.DB_Ontology.prepare('INSERT INTO Complex_Terms VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)')
+			const insert_stmt = env.DB_Ontology.prepare(INSERT_COMPLEX_TERMS_SQL)
 			const insert_stmts = terms.map(({ stem, sense, part_of_speech, structure, pairing, explication, ontology_status, level, notes }) =>
 				insert_stmt.bind(stem, sense, part_of_speech, structure, pairing, explication, ontology_status, level, notes)
 			)
